refactor: Modificados app.jsx y el componente TaskList.jsx [Alvarez Carlos Matias]
