+++ src/components/TaskList.jsx
@@ -1,39 +1,17 @@
-import { useState } from "react";
-//import TaskInput from "./TaskInput";
 import TaskItem from "./TaskItem";
 
-const TaskList = () => {
-  //const [tasks, setTasks] = useState([]);
-  //funcion para añadir
-  //pasar como parametro a TaskInput
-  /*const addTask = (text) => {
-    const newTask = { id: crypto.randomUUID(), text, completed: false };
-    setTasks((prev) => [...prev, newTask]);
-  };*/
-  const toggleTask = (taskId) => {
-    setTasks((prev) =>
-      prev.map((task) =>
-        task.id === taskId ? { ...task, completed: !task.completed } : task
-      )
-    );
-  };
-  const deleteTask = (taskId) => {
-    setTasks((prev) => prev.filter((task) => task.id !== taskId));
-  };
-
+const TaskList = ({ tasks, onToggleTask, onDeleteTask }) => {
   return (
-    <div>
-      <ul>
-        {tasks.map((task) => (
-          <TaskItem
-            key={task.id}
-            task={task}
-            onToggleComplete={toggleTask}
-            onDelete={deleteTask}
-          />
-        ))}
-      </ul>
-    </div>
+    <ul>
+      {tasks.map((task) => (
+        <TaskItem
+          key={task.id}
+          task={task}
+          onToggleComplete={() => onToggleTask(task.id)}
+          onDelete={() => onDeleteTask(task.id)}
+        />
+      ))}
+    </ul>
   );
 };
